[ refactor ] v1 메시징 서비스 패키지 이동 ( #21 )
