+++ src/main/java/com/qring/reservation/application/v1/message/KafkaMessageProducerV1.java
@@ -1,8 +1,8 @@
 package com.qring.reservation.application.v1.message;
 
-import com.qring.reservation.infrastructure.messaging.dto.CreateReservationMessageDTOV1;
-import com.qring.reservation.infrastructure.messaging.dto.SendUserInfoMessageDTOV1;
-import com.qring.reservation.infrastructure.messaging.dto.UpdateReservationMessageDTOV1;
+import com.qring.reservation.infrastructure.messaging.v1.dto.CreateReservationMessageDTOV1;
+import com.qring.reservation.infrastructure.messaging.v1.dto.SendUserInfoMessageDTOV1;
+import com.qring.reservation.infrastructure.messaging.v1.dto.UpdateReservationMessageDTOV1;
 
 public interface KafkaMessageProducerV1 {
