--- src/main/java/com/qring/reservation/infrastructure/messaging/KafkaMessageProducerImplV1.java
@@ -1,31 +1,0 @@
-package com.qring.reservation.infrastructure.messaging;
-
-import com.qring.reservation.application.v1.message.KafkaMessageProducerV1;
-import com.qring.reservation.infrastructure.messaging.dto.CreateReservationMessageDTOV1;
-import com.qring.reservation.infrastructure.messaging.dto.SendUserInfoMessageDTOV1;
-import com.qring.reservation.infrastructure.messaging.dto.UpdateReservationMessageDTOV1;
-import lombok.RequiredArgsConstructor;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class KafkaMessageProducerImplV1 implements KafkaMessageProducerV1 {
-
-    private final KafkaTemplate<String, Object> kafkaTemplate;
-
-    public void publishReservationCreateEvent(CreateReservationMessageDTOV1 message) {
-
-        kafkaTemplate.send("reservation-create-event-topic", message);
-    }
-
-    public void publishReservationUpdateEvent(UpdateReservationMessageDTOV1 message) {
-
-        kafkaTemplate.send("reservation-update-event-topic", message);
-    }
-
-    public void publishUserInfoSendEvent(SendUserInfoMessageDTOV1 message) {
-
-        kafkaTemplate.send("userinfo-send-event-topic", message);
-    }
-}
