+++ src/main/java/com/qring/reservation/infrastructure/messaging/v1/dto/UpdateReservationMessageDTOV1.java
@@ -1,4 +1,4 @@
-package com.qring.reservation.infrastructure.messaging.dto;
+package com.qring.reservation.infrastructure.messaging.v1.dto;
 
 import com.qring.reservation.domain.model.ReservationEntity;
 import lombok.AllArgsConstructor;
